% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildTaipan}
\alias{buildTaipan}
\title{buildTaipan}
\usage{
buildTaipan(questions, images, appdir, launch = TRUE, overwrite = FALSE)
}
\arguments{
\item{questions}{a taipan Questions list of scene and selection questions}

\item{images}{a vector of image locations, can be local or URLs}

\item{appdir}{location to export the completed app}

\item{launch}{launch the app from the new directory after build is completed}

\item{overwrite}{replace the contents of the supplied location with the completed app}
}
\description{
buildTaipan
}
