% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchTaipan}
\alias{launchTaipan}
\title{Launch Taipan}
\usage{
launchTaipan(questions = taipan::sampleQuestions, loadCache = FALSE,
  images = list.files("images"))
}
\arguments{
\item{questions}{A taipan structured list object containing the questions to be asked in the app, see the global environment object sampleQuestions.}

\item{loadCache}{If a cache.Rda object is found in the working directory and loadCache = TRUE, data from previous use of the app will be made avaiable. if there is no object the app will proceed from the beinnning of the image set. The default for this is FALSE.}

\item{images}{The list of images located within the image folder.}
}
\description{
This launches the app
}
\examples{

\dontrun{
library(taipan)

launchTaipan(taipan::tennisQuestionsSmall)
}

}
