% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taipan.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{read_csv}
\alias{reexports}
\alias{write_csv}
\alias{reexports}
\alias{spread}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}}

  \item{readr}{\code{\link[readr]{read_csv}}, \code{\link[readr]{write_csv}}}

  \item{tidyr}{\code{\link[tidyr]{spread}}}
}}

