% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{sampleQuestions}
\alias{sampleQuestions}
\alias{sampleQuestions}
\alias{tennisQuestionsSmall}
\title{Nested list of a sample of Questions}
\format{A nested list with two lists, both containing information for
questions to be displayed:
\itemize{
  \item{scene}{located in the top level list, contains a list of questions
   to be asked about the entire image.}
  \item{selection}{located in the top level list, contains a list of
  questions regarding selected areas within an image.}
  \item{qType}{A character string of the function name of the shiny question
   object to be produced.}
   \item{label}{The question that will be displayed in the shiny app.}
   \item{choices}{A character vector of options to be chosen.}
   \item{selected}{An option argument, when a particular answer should be a
   default selection.}
  ...
}}
\usage{
sampleQuestions

tennisQuestionsSmall
}
\description{
A nested list containing two sets of questions and the information needed
to create shiny objects for each question.

A nested list containing two sets of questions and the information needed
to create shiny objects for each question.
}
\keyword{datasets}
