% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taipan.R
\docType{package}
\name{taipan-package}
\alias{taipan}
\alias{taipan-package}
\title{taipan: Taipan for R}
\description{
Placeholder description
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/srkob1/taipan}
  \item Report bugs at \url{https://github.com/srkob1/taipan/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

}
\keyword{internal}
