% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildTaipan}
\alias{buildTaipan}
\title{buildTaipan}
\usage{
buildTaipan(questions, images, appdir, launch = TRUE, overwrite = FALSE)
}
\arguments{
\item{questions}{a taipan Questions list of scene and selection questions}

\item{images}{a vector of image locations, can be local or URLs}

\item{appdir}{location to export the completed app}

\item{launch}{launch the app from the new directory after build is completed}

\item{overwrite}{replace the contents of the supplied location with the completed app}
}
\description{
buildTaipan
}
\examples{
\dontrun{
library(shiny)
library(taipan)

questions <- taipanQuestions(
  scene = div(radioButtons("graphic", label = ("2D Graphic"),
                           choices = list("Live image", "2D Graphic")),
              radioButtons("bg", label = ("Background"),
                           choices = list("Crowd",
                                          "Court", "Logo wall", "Not applicable")),
              radioButtons("person", label = ("Detectable Person"),
                           choices = list("Yes", "No")),
              radioButtons("shotangle", label = ("Shot angle"),
                           choices = list("Level with players",
                                          "Birds eye",
                                          "Upward angle")),
              radioButtons("situation", label = ("Situation"),
                           choices = list("Court in play",
                                          "Court player close-up",
                                          "Court close-up not player",
                                          "Crowd",
                                          "Off court close up of player",
                                          "Transition"))),
  selection = div(radioButtons("detect", label = ("Detect Face"),
                               choices = list("Player" ,
                                              "Other staff on court", "Fan", "None")),
                  radioButtons("obscured", label = ("Face obscured"),
                               choices = list("Yes", "No")),
                  radioButtons("lighting", label = ("Lighting"),
                               choices = list("Direct sunlight", "Shaded", "Partially shaded")),
                  radioButtons("headangle", label = ("Head angle"),
                               choices = list("Front on", "Back of head",
                                              "Profile", "Other")),
                  radioButtons("glasses", label = ("Glasses"),
                               choices = list("Yes", "No")),
                  radioButtons("visorhat", label = ("Visor/hat"),
                               choices = list("Yes", "No")))
)

buildTaipan(
  questions = questions,
  "https://raw.githubusercontent.com/srkob1/taipan/master/sample_images/2016_CT6_R01_CGarcia_FRA_vs_BStrycova_CZE_WS145_clip.0015.png",
  file.path(tempdir(), "taipan")
)

}

}
