% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchTaipan}
\alias{launchTaipan}
\title{Launch Taipan}
\usage{
launchTaipan(questions = sampleQuestions,
  images = list.files(system.file("images", package = "taipan"), full.names =
  TRUE), answers = NULL)
}
\arguments{
\item{questions}{the questions}

\item{images}{the image folder}

\item{answers}{a dataset of previously created answers from the app}
}
\description{
This launches the app
}
\examples{

\dontrun{
sampleQuestions <- list(scene = list(Q1 = list(qType = "radio",
label = "My First Question",
choices = c("a", "b", "c")),
Q2 = list(qType = "check",
          label = "My Second Question",
          choices = c("1", "2", "3"))),
selection = list(Q1 = list(qType = "check",
                           label = "My First Question",
                           choices = c("a", "b", "c")),
                 Q2 = list(qType = "radio",
                           label = "My Second Question",
                           choices = c("1", "2", "3"))))

launchTaipan(sampleQuestions)
}

}
