% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taipan.R
\docType{package}
\name{taipan-package}
\alias{taipan}
\alias{taipan-package}
\title{taipan: Taipan For R}
\description{
Taipan is a tool for annotating images in preparation for analysis.
It allows users to annotate images in an interactive environment.
Users supply images, questions and the possible answers.
When Images are displayed, survey takers can select the appropriate responses.
The answers are stored in a tidy format for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/srkob1/taipan}
  \item \url{https://github.com/srkob1/taipan/tree/master/inst/images}
  \item Report bugs at \url{https://github.com/srkob1/taipan/issues}
}

}
\author{
\strong{Maintainer}: Stephanie Kobakian \email{stephanie.kobakian@gmail.com}

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}
}

}
\keyword{internal}
